fix: renamed functions-runtime file name
