+++ packages/functions-runtime/src/ModelAPI.js
@@ -4,7 +4,7 @@ const { QueryBuilder } = require("./QueryBuilder");
 const { QueryContext } = require("./QueryContext");
 const { applyWhereConditions } = require("./applyWhereConditions");
 const { applyJoins } = require("./applyJoins");
-const { InlineFile, File } = require("./InlineFile");
+const { InlineFile, File } = require("./File");
 const {
   transformRichDataTypes,
   isPlainObject,
